ARM: sun5i: dt: Add sun5i-a1*-usb-vbus-reg dts include files
Almost all sun5i-a10s boards have a gpio controlled vbus for usb1 using gpio
pin PB10 as in the reference design. This commit adds an include file for this,
avoiding the need to copy and paste the dts code.

For sun5i-a13 based board the reference design uses a gpio on the pmic, which
we don't support yet. Still for consistency with all the other dts files also
add a sun5i-a13-usb-vbus-reg.dtsi files, without a gpio reference for now.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>